python package / lookup function callable from module
