--- test/1147313992.metafits
@@ -1,1 +1,0 @@
-SIMPLE  =                    T / conforms to FITS standard                      BITPIX  =                    8 / array data type                                NAXIS   =                    0 / number of array dimensions                     EXTEND  =                    T                                                  GPSTIME =           1147313992 / [s] GPS time of observation start              EXPOSURE=                  296 / [s] duration of observation                    FILENAME= 'IPS     '           / Name of observation                            MJD     =    57523.09693287037 / [days] MJD of observation                      DATE-OBS= '2016-05-15T02:19:35' / [UT] Date and time of observation             LST     =     24.8144548529336 / [deg] LST                                      HA      = '-23:59:60.00'       / [hours] hour angle of pointing center          AZIMUTH =                  0.0 / [deg] Azimuth of pointing center               ALTITUDE=              76.2838 / [deg] Altitude of pointing center              RA      =     24.6145108681496 / [deg] RA of pointing center                    DEC     =   -13.06836437258106 / [deg] Dec of pointing center                   ATTEN_DB=                  2.0 / [dB] global analogue attenuation, in dB        SUN-DIST=     41.8930721767293 / [deg] Distance from pointing center to Sun     MOONDIST=      137.69369709889 / [deg] Distance from pointing center to Moon    JUP-DIST=     140.672760258617 / [deg] Distance from pointing center to Jupiter GRIDNAME= 'sweet   '           / Pointing grid name                             GRIDNUM =                    9 / Pointing grid number                           CREATOR = 'jmorgan '           / Observation creator                            PROJECT = 'D0011   '           / Project ID                                     MODE    = 'HW_LFILES'          / Observation mode                               RECVRS  = '1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16' / Active receivers           DELAYS  = '6,6,6,6,4,4,4,4,2,2,2,2,0,0,0,0' / Beamformer delays                 CALIBRAT=                    F / Intended for calibration                       CENTCHAN=                  121 / Center coarse channel                          CHANNELS= '57,58,59,60,61,62,63,64,65,66,67,68,121,122,123,124,125,126,127,128&'CONTINUE  ',129,130,131,132&'                                                   CONTINUE  '' / Coarse channels                                                  CHANSEL = '0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23' / IndiSUN-ALT =     37.1191821388112 / [deg] Altitude of Sun                          FINECHAN=                   40 / [kHz] Fine channel width - correlator freq_res INTTIME =                  0.5 / [s] Individual integration time                NAV_FREQ=                    4 / Assumed frequency averaging                    NSCANS  =                  592 / Number of time instants in correlation productsNINPUTS =                  256 / Number of inputs into the correlation products NCHANS  =                  768 / Number of (averaged) fine channels in spectrum BANDWDTH=                30.72 / [MHz] Total bandwidth                          FREQCENT=              154.255 / [MHz] Center frequency of observation          TIMEOFF =                    0 / [s] Deprecated, use QUACKTIM or GOODTIME       DATESTRT= '2016-05-15T02:19:35' / [UT] Date and time of correlations start      VERSION =                  2.0 / METAFITS version number                        TELESCOP= 'MWA     '                                                            INSTRUME= '128T    '                                                            QUACKTIM=                  0.5 / Seconds of bad data after observation starts   GOODTIME=         1463278775.5 / OBSID+QUACKTIME as Unix timestamp              DATE    = '2019-07-05T05:46:00' / UT Date of file creation                      COMMENT After an observation starts, the receiver hardware changes take a       COMMENT few seconds to stabilise. The 24 GPU boxes start saving data to         COMMENT their output files anywhere from a second _before_ the start of         COMMENT the observation to a few seconds after, and not necessarily at the      COMMENT same time on each gpubox.                                               COMMENT QUACKTIM and GOODTIME represent the start of the first uncontaminated   COMMENT data, rounded up to the next time-averaged data packet. Note that this  COMMENT time may be before the first actual data in some or all gpubox files.   HISTORY Created by user apache                                                  HISTORY Created on host mwa-django01p.curtin.edu.au                             HISTORY Command: "mod_wsgi"                                                     END
